src/test/java/_0851_0900/_853_Car_Fleet/SolutionITest.java
package _0851_0900._853_Car_Fleet;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 3/28/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase1(Solution solution) {
    final int target = 12, position[] = {10, 8, 0, 5, 3}, speed[] = {2, 4, 1, 1, 3};
    assertEquals(3, solution.carFleet(target, position, speed));
  }

  static Stream<Solution> solutionProvider() {
    return Stream.of(
        new SolutionI()
    );
  }
}